# doc/a3s-openapi.yaml
openapi: 3.0.0
info:
  description: >-
    API Definition for the A3S. This service allows authentication, authorisation and accounting.
  version: 1.0.1
  title: A3S
paths:
...
  '/functions':
    get:
      tags:
        - Function
      description: Search for functions.
      summary: Search for functions.
      operationId: listFunctions
      parameters:
        - name: permissions
          in: query
          required: false
          description: |
            If this field is set, then the permission list is filled in
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          required: false
          description: The page to view.
          schema:
            type: integer
            minimum: 0
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          description: The size of a page.
          schema:
            type: integer
            minimum: 1
            maximum: 20
            format: int32
            default: 10
        - name: filterDescription
          in: query
          required: false
          description: A search query filter on the description
          schema:
            type: string
            minLength: 0
            maxLength: 255
          example: FilterByEntityType
        - name: orderBy
          in: query
          required: false
          style: form
          explode: false
          description: >-
            a comma separated list of fields in their sort order. Ascending
            order is assumed. Append desc after a field to indicate descending
            order.
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          headers:
            X-Pagination:
              description: Pagination header.
              schema:
                $ref: '#/components/schemas/PaginationHeaderResponse'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Function'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to access the list of functions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Function list not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Function
      summary: Create a new function.
      description: Create a new function.
      operationId: createFunction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FunctionSubmit'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Function'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to create functions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Function related entity (such as permissions) not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/functions/{functionId}':
    get:
      tags:
        - Function
      summary: Get a function.
      description: Get a function by its UUID.
      operationId: getFunction
      parameters:
        - name: functionId
          in: path
          description: function
          required: true
          schema:
            $ref: '#/components/schemas/FunctionId'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Function'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to access the function.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Function not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Function
      summary: Update a function.
      description: Update a function by its UUID.
      operationId: updateFunction
      parameters:
        - name: functionId
          in: path
          description: The UUID of the function.
          required: true
          schema:
            $ref: '#/components/schemas/FunctionId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FunctionSubmit'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Function'
        '204':
          description: No Content.
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to update the function.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Functions not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Function
      summary: Delete a function.
      description: Deletes a function from A3S.
      operationId: deleteFunction
      parameters:
        - name: functionId
          description: The UUID of the function to delete.
          required: true
          in: path
          schema:
            $ref: '#/components/schemas/FunctionId'
      responses:
        '204':
          description: No Content.
        '400':
          description: Invalid parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not authorized to delete functions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Function not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  # Endregion - Functions
  
  # Region  - ApplicationFunctions
...
components:
  schemas:
...
    ApplicationId:
      type: string
      description: The UUID of an application.
      format: uuid
...
    PermissionId:
      type: string
      format: uuid
    Permission:
      description: |
        Permission for an application
      type: object
      properties:
        uuid:
          $ref: '#/components/schemas/PermissionId'
        name:
          type: string
        description:
          type: string

    # End Region - Permissions      
          
    # Region - Functions
    FunctionId:
      type: string
      format: uuid
    Function:
      description: |
        A Function - A collection of permissions to denote a piece of functionality in an application. Must be assigned to an application.
      type: object
      properties:
        uuid:
          $ref: '#/components/schemas/FunctionId'
        name:
          type: string
        description:
          type: string
        applicationId:
          $ref: '#/components/schemas/ApplicationId'
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
    FunctionSubmit:
      description: |
        Model used to create an new function or update an existing Function.
      required:
        - uuid
        - name
        - description
        - applicationId
        - permissions
      type: object
      properties:
        uuid:
          $ref: '#/components/schemas/FunctionId'
        name:
          type: string
          description: The name of the function.
        description:
          description: A brief description of the function.
          type: string
        applicationId:
          $ref: '#/components/schemas/ApplicationId'
        permissions:
          description: A list of permission UUIDs that are to be added to the function.
          type: array
          items:
            type: string
            format: uuid
    # End Region - Functions
    
    # Region - Application Function
...
    ErrorResponse:
      description: Error response object to give more info with HTTP status code
      type: object
      required:
        - message
      properties:
        message:
          type: string
    PaginationHeaderResponse:
      description: >-
        This is the standard response for requests that have paginated
        collections
      type: object
      properties:
        total:
          type: integer
          format: int32
          minimum: 1
          description: The total number of pages in the result set
        count:
          type: integer
          format: int32
          minimum: 1
          description: The total number of results in the result set
        size:
          type: integer
          format: int32
          minimum: 1
          maximum: 20
          default: 10
          description: >-
            The number items to include in a page of results. The page size for
            the result set
        current:
          type: integer
          format: int32
          minimum: 1
          description: >-
            The position of the page in the paged result set that is being
            returned
        first:
          type: string
          format: url
          description: >-
            The link to the first page of results containing [size] results.
            This link includes the page, size, filter, orderBy and fields query
            parameters.
        last:
          type: string
          format: url
          description: >-
            The link to the last page of results containing [size] results. This
            link includes the page, size, filter, orderBy and fields query
            parameters. If this is the first page in the result set then this
            will be [null].
        prev:
          type: string
          format: url
          description: >-
            The link to the previous page of results containing [size] results.
            This link includes the page, size, filter, orderBy and fields query
            parameters. If this is the first page in the result set then this
            will be [null].
        next:
          type: string
          format: url
          description: >-
            The link to the next page of results containing [size] results. This
            link includes the page, size, filter, orderBy and fields query
            parameters. If this is the last page in the result set then this will
            be [null].


    #endregion